pyproject.toml
[project]
name = "overtime-report"
version = "0.1.0"
description = "A CLI tool to generate development work reports from Git history using AI."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "json5>=0.9.0",
    "socksio>=1.0.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/overtime_report"]
